Add Permission Wilayah

+++ app/Http/Controllers/Api/MasterData/WilayahController.php
@@ -7,49 +7,63 @@ use Illuminate\Http\Request;
 use App\Models\MasterData\Wilayah;
 use App\Http\Resources\MasterData\WilayahResponse;
 use App\Http\Requests\MasterData\WilayahRequest;
+use App\Http\Controllers\ApiLogController as ApiLog;
+
 class WilayahController extends Controller
 {
+    protected $apiLog;
+    public function __construct()
+    {
+        $this->apiLog           = new ApiLog;
+    }
     public function index(Request $request)
     {
-        $data = Wilayah::query();
-        $per_page = 10;
-        if($request->filled("per_page")){
-            $per_page = $request->per_page;
-        }
-        if($request->sort_field && $request->sort_type){
-            $data = $data->orderBy($request->sort_field,$request->sort_type);
-        }
-        if($request->filled('rt')){
-            $data =  $data->where("rt","LIKE","%".$request->rt."%");
-        }
-        if($request->filled('rw')){
-            $data =  $data->where("rw","LIKE","%".$request->rw."%");
-        }
-        if($request->filled('kelurahan')){
-            $data =  $data->where("kelurahan","LIKE","%".$request->kelurahan."%");
-        }
-        if($request->filled('kecamatan')){
-            $data =  $data->where("kecamatan","LIKE","%".$request->kecamatan."%");
-        }
-        if($request->filled('kabupaten')){
-            $data =  $data->where("kabupaten","LIKE","%".$request->kabupaten."%");
-        }
-        if($request->filled('provinsi')){
-            $data =  $data->where("provinsi","LIKE","%".$request->provinsi."%");
-        }
-        if($request->filled('negara')){
-            $data =  $data->where("negara","LIKE","%".$request->negara."%");
-        }
-        if($request->filled('kode_pos')){
-            $data =  $data->where("kode_pos","LIKE","%".$request->kode_pos."%");
-        }
-        if($request->filled('jalan')){
-            $data =  $data->where("jalan","LIKE","%".$request->jalan."%");
-        }
-        $data = $data->paginate($per_page);
-        return (WilayahResponse::collection($data))
+        $checkPermission    = $this->apiLog->checkPermission(class_basename(get_class($this)), 'pread');
+        if ($checkPermission) {
+            $data = Wilayah::query();
+            $per_page = 10;
+            if ($request->filled("per_page")) {
+                $per_page = $request->per_page;
+            }
+            if ($request->sort_field && $request->sort_type) {
+                $data = $data->orderBy($request->sort_field, $request->sort_type);
+            }
+            if ($request->filled('rt')) {
+                $data =  $data->where("rt", "LIKE", "%" . $request->rt . "%");
+            }
+            if ($request->filled('rw')) {
+                $data =  $data->where("rw", "LIKE", "%" . $request->rw . "%");
+            }
+            if ($request->filled('kelurahan')) {
+                $data =  $data->where("kelurahan", "LIKE", "%" . $request->kelurahan . "%");
+            }
+            if ($request->filled('kecamatan')) {
+                $data =  $data->where("kecamatan", "LIKE", "%" . $request->kecamatan . "%");
+            }
+            if ($request->filled('kabupaten')) {
+                $data =  $data->where("kabupaten", "LIKE", "%" . $request->kabupaten . "%");
+            }
+            if ($request->filled('provinsi')) {
+                $data =  $data->where("provinsi", "LIKE", "%" . $request->provinsi . "%");
+            }
+            if ($request->filled('negara')) {
+                $data =  $data->where("negara", "LIKE", "%" . $request->negara . "%");
+            }
+            if ($request->filled('kode_pos')) {
+                $data =  $data->where("kode_pos", "LIKE", "%" . $request->kode_pos . "%");
+            }
+            if ($request->filled('jalan')) {
+                $data =  $data->where("jalan", "LIKE", "%" . $request->jalan . "%");
+            }
+            $data = $data->paginate($per_page);
+            return (WilayahResponse::collection($data))
                 ->response()
                 ->setStatusCode(200);
+        } else {
+            return response()->json([
+                'error' => 'Forbidden access'
+            ], 403);
+        }
     }
     /**
      * Store a newly created resource in storage.
@@ -59,11 +73,18 @@ class WilayahController extends Controller
      */
     public function store(WilayahRequest $request)
     {
-        $validate = $request->validated();
-        $data = Wilayah::create($validate);
-        return (New WilayahResponse($data))
+        $checkPermission    = $this->apiLog->checkPermission(class_basename(get_class($this)), 'pcreate');
+        if ($checkPermission) {
+            $validate = $request->validated();
+            $data = Wilayah::create($validate);
+            return (new WilayahResponse($data))
                 ->response()
                 ->setStatusCode(201);
+        } else {
+            return response()->json([
+                'error' => 'Forbidden access'
+            ], 403);
+        }
     }
 
     /**
@@ -74,10 +95,17 @@ class WilayahController extends Controller
      */
     public function show($id)
     {
-        $data = Wilayah::findOrFail($id);
-        return (New WilayahResponse($data))
+        $checkPermission    = $this->apiLog->checkPermission(class_basename(get_class($this)), 'pread');
+        if ($checkPermission) {
+            $data = Wilayah::findOrFail($id);
+            return (new WilayahResponse($data))
                 ->response()
                 ->setStatusCode(200);
+        } else {
+            return response()->json([
+                'error' => 'Forbidden access'
+            ], 403);
+        }
     }
 
     /**
@@ -91,15 +119,20 @@ class WilayahController extends Controller
     {
         // return $request->all();
 
+        $checkPermission    = $this->apiLog->checkPermission(class_basename(get_class($this)), 'pread');
+        if ($checkPermission) {
+            $validate = $request->validated();
 
-        $validate = $request->validated();
-
-        $data = Wilayah::findOrFail($id);
-        if($data->update($validate)){
-            return (New WilayahResponse($data))
-                ->response()
-                ->setStatusCode(200);
-
+            $data = Wilayah::findOrFail($id);
+            if ($data->update($validate)) {
+                return (new WilayahResponse($data))
+                    ->response()
+                    ->setStatusCode(200);
+            }
+        } else {
+            return response()->json([
+                'error' => 'Forbidden access'
+            ], 403);
         }
     }
 
@@ -111,11 +144,18 @@ class WilayahController extends Controller
      */
     public function destroy($id)
     {
-        $data = Wilayah::findOrFail($id);
-        if($data->delete()){
-            return (New WilayahResponse($data))
-                ->response()
-                ->setStatusCode(200);
+        $checkPermission    = $this->apiLog->checkPermission(class_basename(get_class($this)), 'pdelete');
+        if ($checkPermission) {
+            $data = Wilayah::findOrFail($id);
+            if ($data->delete()) {
+                return (new WilayahResponse($data))
+                    ->response()
+                    ->setStatusCode(200);
+            }
+        } else {
+            return response()->json([
+                'error' => 'Forbidden access'
+            ], 403);
         }
     }
 }
